+++ src/context/dataActions.js
@@ -8,6 +8,7 @@ import { createPartnerActions } from './actions/partners';
 import { createCatalogueActions } from './actions/catalogue';
 import { createKitActions } from './actions/kits';
 import { createInverterActions } from './actions/inverters';
+import { createPompeKitActions } from './actions/pompeKits';
 import { createPayoutActions } from './actions/payouts';
 import { createDevisActions } from './actions/devis';
 import { createFormationActions } from './actions/formations';
@@ -23,6 +24,7 @@ export function createActions(setState) {
     ...createCatalogueActions(setState),
     ...createKitActions(setState),
     ...createInverterActions(setState),
+    ...createPompeKitActions(setState),
     ...createPayoutActions(setState),
     ...createDevisActions(setState),
     ...createFormationActions(setState),